Code is now VS2015 compliant, and 5.7 Qt works as well. Font Renderer stalls when adding a new layer now though.

+++ Harmony/include/Harmony/Renderer/OpenGL/HyOpenGLShader.h
@@ -37,7 +37,7 @@ public:
 	void CompileFromString(HyShaderType eType);
 	void Link();
 	void Use();
-	void SetVertexAttributePtrs(uint32 uiStartOffset);
+	void SetVertexAttributePtrs(size_t uiStartOffset);
 
 	int GetHandle();
 	bool IsLinked();
